refactor: move cli helpers into cli_util module
We cannot name the file "util.py" and "from uaclient.cli import util as
cli_util" because there is a bug in python where python will confuse the
two modules even if one is renamed at import time.
