+++ uaclient/timer/metering.py
@@ -4,11 +4,11 @@ Functions to be used when running metering jobs
 
 from uaclient import config
 from uaclient.api.u.pro.status.is_attached.v1 import _is_attached
-from uaclient.cli import assert_lock_file
+from uaclient.cli import cli_util
 from uaclient.contract import UAContractClient
 
 
-@assert_lock_file("timer metering job")
+@cli_util.assert_lock_file("timer metering job")
 def metering_enabled_resources(cfg: config.UAConfig) -> bool:
     # We only run this job if there is no other job running.
     # The reason for that is to avoid potential conflicts with
